refactor(k8s): split the warm Job manifest tests into a sibling file
`warm_job.rs` was already at 49.9 KiB against the Server Guards 51200-byte
limit, and the two projected lease-identity constants pushed it to 51357.

The test module moves verbatim into `warm_job_tests.rs`, included with `#[path]`
— the same convention `graph_warmer_ledger_tests.rs` already uses — so it stays a
child of the module under test and keeps its `use super::*` access. No
production logic is touched and no `djinn:allow-oversize` marker is added.

  warm_job.rs  51357 -> 25449 bytes  (+ warm_job_tests.rs 24006)
